Move libi8080 to root directory
